+++ forge/src/test/java/tests/EnergyTest.java
@@ -96,7 +96,7 @@ public class EnergyTest {
         }
 
         @Override
-        public int getVoltage() {
+        public long getVoltage() {
             return voltage;
         }
